eighth C Program

+++ pointers_arrays_strings/7-leet.c
@@ -19,7 +19,7 @@ char num[5] = {'4', '3', '0', '7', '1'};
  for (i = 0; c[i] != '\0'; i++)
    {
   int j;
-  int k = 0;
+  float k = 0;
   for (j = 0; let[j] != '\0'; j++)
    {
      if (c[i] == let[j])
